move FilterNames to plugin
