+++ yellowstone-grpc-geyser/Cargo.toml
@@ -45,7 +45,7 @@ tokio = { workspace = true, features = ["rt-multi-thread", "macros", "fs"] }
 tokio-stream = { workspace = true }
 tonic = { workspace = true, features = ["gzip", "zstd", "tls", "tls-roots"] }
 tonic-health = { workspace = true }
-yellowstone-grpc-proto = { workspace = true, features = ["convert"] }
+yellowstone-grpc-proto = { workspace = true, features = ["convert", "plugin"] }
 
 [build-dependencies]
 anyhow = { workspace = true }
