+++ yellowstone-grpc-proto/src/plugin/mod.rs
@@ -1,0 +1,1 @@
+pub mod filter;
